+++ .gitignore
@@ -4,6 +4,7 @@
 .vscode/*
 *.iml
 node_modules
+package-lock.json
 
 
 # Google Drive exclude
