+++ features/projects/detail/detail.html
@@ -33,7 +33,7 @@
         <div class="collapse navbar-collapse"
              id="navbarCollapse">
             <div class="navbar-nav">
-                <a href="/"
+                <a href="/projects"
                    class="nav-item nav-link">Projekte</a>
             </div>
         </div>
@@ -43,7 +43,7 @@
     <div class="container">
         <ul>
             <li><a href="/">FH JOANNEUM</a></li>
-            <li><a href="/">PROJEKTE</a></li>
+            <li><a href="/projects">PROJEKTE</a></li>
             <li><a id="breadcrumb" href=""></a></li>
         </ul>
     </div>
