+++ features/projects/list.html
@@ -40,7 +40,7 @@
             <div class="collapse navbar-collapse"
                  id="navbarCollapse">
                 <div class="navbar-nav">
-                    <a href="/"
+                    <a href="/projects"
                        class="nav-item nav-link">Projekte</a>
                 </div>
             </div>
@@ -51,7 +51,7 @@
         <div class="container">
             <ul>
                 <li><a href="/">FH JOANNEUM</a></li>
-                <li><a href="/">PROJEKTE</a></li>
+                <li><a href="/projects">PROJEKTE</a></li>
             </ul>
         </div>
     </div>
