+++ features/projects/projectService.js
@@ -1,0 +1,17 @@
+
+module.exports = class ProjectService {
+    constructor() {
+    }
+  
+    get() {
+        return [];
+    }
+    
+    find(id) {
+        return {};
+    }
+    
+    insert(project) {
+        return "";
+    }
+  };
