+++ package.json
@@ -21,6 +21,7 @@
   },
   "homepage": "https://github.com/SnickersVsMars/webapp-ws20-projects#readme",
   "dependencies": {
+    "config": "^3.3.2",
     "express": "^4.17.1"
   }
 }
