+++ public/js/detail.js
@@ -21,7 +21,7 @@ $.ajax({
                 document.getElementById("customer").innerHTML=validate(project.customer);
                 document.getElementById("costCenter").innerHTML=validate(project.costCenter);
                 document.getElementById("breadcrumb").innerHTML=`PROJEKT ${project.number}`;
-                document.getElementById("breadcrumb").setAttribute('href', '/project/'+project.id);
+                document.getElementById("breadcrumb").setAttribute('href', '/projects/'+project.id);
 
                 fillEmployess(project.employees);
                 fillMilestones(project.milestones);
